file upload service passed through context to the resolver
