+++ src/apollo-server.ts
@@ -1,5 +1,6 @@
 import { ApolloServer } from 'apollo-server-express';
 import { schema } from './api/graphql/schema';
+import { FileUploadService } from './services/file-upload-service';
 
 export const apolloServer = new ApolloServer({
   schema,
@@ -11,5 +12,8 @@ export const apolloServer = new ApolloServer({
     maxFileSize: 5 * 1024 * 1024, // 5 MB
     maxFiles: 1
   },
-  introspection: true
+  introspection: true,
+  context: {
+    fileUploadService: new FileUploadService()
+  }
 });
