+++ src/services/file-upload-service/index.ts
@@ -1,0 +1,2 @@
+export { FileUploadService } from './file-upload-service';
+export { FileUploadResponse } from './file-upload-service.interface';
